Use uint64 for indirect pin refcounts
Platform-dependent behavior is not nice, and negative refcounts are
not very useful.

License: MIT

+++ core/commands/pin.go
@@ -274,7 +274,7 @@ Defaults to "direct".
 
 type RefKeyObject struct {
 	Type  string
-	Count int
+	Count uint64
 }
 
 type RefKeyList struct {
